backend/requirements.txt
bcrypt
flask
flask-emails
flask-httpauth
flask-marshmallow
flask-marshmallow-openapi==0.6.4
flask-migrate
flask-sqlalchemy
gevent
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
